+++ projects/igniteui-angular/src/lib/grids/grid/row-drag.directive.spec.ts
@@ -1,4 +1,4 @@
-import { Component, ViewChild, DebugElement, QueryList } from '@angular/core';
+import { Component, ViewChild, DebugElement, QueryList, TemplateRef } from '@angular/core';
 import { TestBed, ComponentFixture, fakeAsync, tick } from '@angular/core/testing';
 import { FormsModule } from '@angular/forms';
 import { By } from '@angular/platform-browser';
@@ -449,6 +449,40 @@ describe('Row Drag Tests #grid', () => {
 
                 const ghostText = document.getElementsByClassName(CSS_CLASS_GHOST_ROW)[0].textContent;
                 expect(ghostText).toEqual(' Moving a row! ');
+            });
+
+            it('should allow setting custom drag icon and ghost element via Input.', () => {
+                fixture = TestBed.createComponent(IgxGridRowCustomGhostDraggableComponent);
+                grid = fixture.componentInstance.instance as IgxGridComponent;
+                fixture.detectChanges();
+                dropAreaElement = fixture.debugElement.query(By.css(CSS_CLASS_DROPPABLE_AREA)).nativeElement;
+                grid.dragIndicatorIconTemplate = fixture.componentInstance.rowDragTemplate;
+                grid.dragGhostCustomTemplate = fixture.componentInstance.rowDragGhostTemplate;
+                fixture.detectChanges();
+                rows = grid.rowList.toArray();
+                dragIndicatorElements = fixture.debugElement.queryAll(By.css(CSS_CLASS_DRAG_INDICATOR));
+                dragIndicatorElement = dragIndicatorElements[2].nativeElement;
+                dragRows = fixture.debugElement.queryAll(By.directive(IgxRowDragDirective));
+                rowDragDirective = dragRows[1].injector.get(IgxRowDragDirective);
+
+                expect(dragIndicatorElement.textContent.trim()).toBe('expand_less');
+
+                startPoint = UIInteractions.getPointFromElement(dragIndicatorElement);
+                movePoint = UIInteractions.getPointFromElement(rows[4].nativeElement);
+                dropPoint = UIInteractions.getPointFromElement(dropAreaElement);
+                pointerDownEvent = UIInteractions.createPointerEvent('pointerdown', startPoint);
+                pointerMoveEvent = UIInteractions.createPointerEvent('pointermove', movePoint);
+
+                rowDragDirective.onPointerDown(pointerDownEvent);
+                rowDragDirective.onPointerMove(pointerMoveEvent);
+                pointerMoveEvent = UIInteractions.createPointerEvent('pointermove', dropPoint);
+                rowDragDirective.onPointerMove(pointerMoveEvent);
+                const ghostElements: HTMLCollection = document.getElementsByClassName(CSS_CLASS_GHOST_ROW);
+                expect(ghostElements.length).toEqual(1);
+
+                const ghostText = document.getElementsByClassName(CSS_CLASS_GHOST_ROW)[0].textContent;
+                expect(ghostText.trim()).toEqual('CUSTOM');
+
             });
         });
     });
@@ -1228,11 +1262,26 @@ export class IgxGridRowDraggableComponent extends DataParent {
         <div #nonDroppableArea class="non-droppable-area"
         [ngStyle]="{width:'100px', height:'100px', backgroundColor:'yellow'}">
         </div>
+
+        <ng-template #rowDragGhostTemplate let-data igxRowDragGhost>
+                <div class="dragGhost">
+                        CUSTOM
+                </div>
+        </ng-template>
+        <ng-template #rowDragTemplate let-data igxDragIndicatorIcon>
+            <igx-icon>expand_less</igx-icon>
+        </ng-template>
     `
 })
 export class IgxGridRowCustomGhostDraggableComponent extends DataParent {
     @ViewChild(IgxGridComponent, { read: IgxGridComponent, static: true })
     public instance: IgxGridComponent;
+
+    @ViewChild('rowDragGhostTemplate', {read: TemplateRef, static: true })
+    public rowDragGhostTemplate: TemplateRef<any>;
+
+    @ViewChild('rowDragTemplate', {read: TemplateRef, static: true })
+    public rowDragTemplate: TemplateRef<any>;
 
     @ViewChild('dropArea', { read: IgxDropDirective, static: true })
     public dropArea: IgxDropDirective;